changing up the byte structure

+++ src/main/java/model/Operation.java
@@ -3,21 +3,21 @@ package model;
 /**
  * Created by nathaniel on 8/29/16.
  */
-public enum Operation{
+public enum Operation {
 
-    GETTIME(0),
-    CHANGETIME(1);
+    GETTIME((short) 0),
+    CHANGETIME((short) 1);
 
-    public final int id;
+    public final short id;
 
-    Operation(int id) {
+    Operation(short id) {
         this.id = id;
     }
 
-    public static Operation getOperation(int id){
+    public static Operation getOperation(short id) {
         Operation[] ops = Operation.values();
-        for(int i = 0; i < ops.length; ++i){
-            if(ops[i].id == id)
+        for (int i = 0; i < ops.length; ++i) {
+            if (ops[i].id == id)
                 return ops[i];
         }
         throw new RuntimeException("No Such Operations");
